backend/runner/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-blinka-displayio==2.1.2
adafruit-circuitpython-bitmap_font==2.1.3
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-display-text==3.2.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ticks==1.1.1
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.76.1
Adafruit-PureIO==1.1.11
atomics==1.0.3
binho-host-adapter==0.1.6
cffi==1.17.1
pillow==11.1.0
pycparser==2.22
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
sysv_ipc==1.1.0
typing_extensions==4.12.2
